2020_10_07 Props追加、マップ読み込み方式変更(画像から)

+++ Animation.cpp
@@ -7,6 +7,7 @@
 
 void Animation::Init() {
 
+	// データ初期化
 	mState = "Idle";
 	mNewState = "Idle";
 	mFrame = 0;
@@ -18,10 +19,10 @@ void Animation::Uninit() {
 
 void Animation::Update() {
 
-	// Frame 処理
+	// フレーム処理
 	mFrame += mCoefficient * mAnimationSpeed * Time::GetDeltaTime();
 
-	// Blend 処理
+	// ブレンド処理
 	if (mState != mNewState) {
 		mBlend += 3.0f * Time::GetDeltaTime();
 	}
@@ -31,6 +32,7 @@ void Animation::Update() {
 		mBlend = 0.0f;
 	}
 
+	// ForDebug
 	if (GetUsePanel()) {
 		DataPanel();
 	}
@@ -39,14 +41,13 @@ void Animation::Update() {
 
 void Animation::FixedUpdate() {
 
+	// フレーム修正
 	if (mFrame >= 1000000) {
 		mFrame = 0;
 	}
-
 	if (mBlend < 0.0f) {
 		mBlend = 0.0f;
 	}
-
 	if (mBlend >= 1.0f) {
 		mBlend = 1.0f;
 	}
@@ -58,7 +59,7 @@ void Animation::DataPanel() {
 	ImGui::Begin(GetResource()->Name.c_str());
 	if (ImGui::TreeNode(u8"アニメーション")) {
 		ImGui::Text("Frame : %f", mFrame);
-		ImGui::Text(u8"再生速度 : %f", (float)mCoefficient * (float)mAnimationSpeed);
+		ImGui::Text(u8"再生速度 : %f", mCoefficient * mAnimationSpeed);
 		ImGui::Text(u8"Blend : %f", mBlend);
 		ImGui::Text(u8"状態 : %s", mState.c_str());
 		ImGui::Text(u8"状態(新) : %s", mNewState.c_str());
